refactor(web): group hooks/ into workspace/, artifacts/, analysis/ sub-dirs
Organises 14 flat hook files into three concern-based sub-directories:
  hooks/workspace/ — preferences, URL state, theme
  hooks/artifacts/ — artifact loading, reload, polling, page orchestration
  hooks/analysis/  — omnibox search, runs results, resource code

Updates all ~24 import sites (external + internal relative paths).
